add test for canceling subscriptions
